[Fix] Fix various sprintf("%d") where the argument was in fact size_t.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@6706 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/PatternEditorDialogs.cpp
@@ -139,7 +139,7 @@ BOOL CPatternPropertiesDlg::OnInitDialog()
 			m_nPattern,
 			pattern.GetNumRows(),
 			(pattern.GetNumRows() == 1) ? _T("") : _T("s"),
-			(pattern.GetNumRows() * sndFile.GetNumChannels() * sizeof(ModCommand)) / 1024);
+			static_cast<int>((pattern.GetNumRows() * sndFile.GetNumChannels() * sizeof(ModCommand)) / 1024));
 		SetDlgItemText(IDC_TEXT1, s);
 
 		// Window title
